backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
google-genai==1.14.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pytest==8.3.4
pytest-asyncio==0.25.0
